+++ src/Adapter/AssertType/ASSERT_TYPE.mjs
@@ -1,7 +1,0 @@
-export const ASSERT_TYPE_CSS = "css";
-
-export const ASSERT_TYPE_JSON = "json";
-
-/**
- * @typedef {ASSERT_TYPE_CSS | ASSERT_TYPE_JSON} ASSERT_TYPE
- */