+++ src/Adapter/Cookie/SET_COOKIE_OPTION.mjs
@@ -1,0 +1,15 @@
+export const SET_COOKIE_OPTION_DOMAIN = "Domain";
+
+export const SET_COOKIE_OPTION_EXPIRES = "Expires";
+
+export const SET_COOKIE_OPTION_HTTP_ONLY = "HttpOnly";
+
+export const SET_COOKIE_OPTION_MAX_AGE = "Max-Age";
+
+export const SET_COOKIE_OPTION_PATH = "Path";
+
+export const SET_COOKIE_OPTION_PRIORITY = "Priority";
+
+export const SET_COOKIE_OPTION_SAME_SITE = "SameSite";
+
+export const SET_COOKIE_OPTION_SECURE = "Secure";