+++ src/Adapter/Cookie/SET_COOKIE_SAME_SITE.mjs
@@ -1,0 +1,5 @@
+export const SET_COOKIE_SAME_SITE_LAX = "Lax";
+
+export const SET_COOKIE_SAME_SITE_NONE = "None";
+
+export const SET_COOKIE_SAME_SITE_STRICT = "Strict";