+++ src/Adapter/Fetch/FETCH_ASSERT_TYPE.mjs
@@ -1,0 +1,3 @@
+export const FETCH_ASSERT_TYPE_CSS = "css";
+
+export const FETCH_ASSERT_TYPE_JSON = "json";