+++ src/Adapter/Fetch/Fetch.mjs
@@ -1,6 +1,3 @@
-/** @typedef {import("../AssertType/ASSERT_TYPE.mjs").ASSERT_TYPE} ASSERT_TYPE */
-/** @typedef {import("../Method/METHOD.mjs").METHOD} METHOD */
-
 /**
- * @typedef {{url: string, query_params?: {[key: string]: *} | null, method?: METHOD, data?: *, abort_signal?: AbortSignal, no_ui?: boolean, no_error_ui?: boolean, no_authenticate?: boolean, assert_type?: ASSERT_TYPE}} Fetch
+ * @typedef {{url: string, query_params?: {[key: string]: *} | null, method?: string, data?: *, abort_signal?: AbortSignal, no_ui?: boolean, no_error_ui?: boolean, no_authenticate?: boolean, assert_type?: string}} Fetch
  */