+++ src/Adapter/Fetch/showError.mjs
@@ -1,3 +1,0 @@
-/**
- * @typedef {(error: Error | (Response & {ok: false})) => Promise<boolean>} showError
- */